+++ apps/rhf-mui-docs/src/page-components/mui/select/RHFMultiSelectDropdown.tsx
@@ -1,0 +1,29 @@
+import MarkdownTable from '@site/src/components/markdown-table';
+import { PropsDescription } from '@site/src/constants';
+
+const RHFMultiSelectDropdownPropsTable = () => {
+  const tableRows = [
+    PropsDescription.fieldName,
+    PropsDescription.control,
+    PropsDescription.registerOptions,
+		PropsDescription.options,
+		PropsDescription.labelKey,
+		PropsDescription.valueKey,
+    PropsDescription.onValueChange_CountrySelect,
+		PropsDescription.label,
+    PropsDescription.showLabelAboveFormField,
+    PropsDescription.formLabelProps,
+		PropsDescription.checkboxProps,
+		PropsDescription.formControlLabelProps,
+    PropsDescription.helperText,
+    PropsDescription.errorMessage,
+    PropsDescription.hideErrorMessage,
+    PropsDescription.formHelperTextProps
+  ];
+
+  return (
+    <MarkdownTable rows={tableRows} showType/>
+  );
+};
+
+export default RHFMultiSelectDropdownPropsTable;