+++ apps/rhf-mui-docs/src/page-components/mui/select/index.ts
@@ -1,3 +1,4 @@
 export { default as RHFCountrySelectPropsTable } from './RHFCountrySelect';
+export { default as RHFMultiSelectDropdownPropsTable } from './RHFMultiSelectDropdown';
 export { default as RHFNativeSelectPropsTable } from './RHFNativeSelect';
 export { default as RHFSelectPropsTable } from './RHFSelect';